Issue #844 Move StoreConfigurationImpl to internal
