+++ impl/src/test/java/org/ehcache/impl/internal/store/offheap/OffHeapStoreSPITest.java
@@ -19,7 +19,7 @@ package org.ehcache.impl.internal.store.offheap;
 import org.ehcache.config.EvictionVeto;
 import org.ehcache.config.ResourcePool;
 import org.ehcache.config.ResourcePools;
-import org.ehcache.core.config.store.StoreConfigurationImpl;
+import org.ehcache.core.internal.store.StoreConfigurationImpl;
 import org.ehcache.config.builders.ResourcePoolsBuilder;
 import org.ehcache.config.units.MemoryUnit;
 import org.ehcache.expiry.Expirations;
