+++ impl/src/test/java/org/ehcache/impl/internal/store/offheap/OffHeapStoreTest.java
@@ -18,7 +18,7 @@ package org.ehcache.impl.internal.store.offheap;
 
 import org.ehcache.config.EvictionVeto;
 import org.ehcache.config.ResourceType;
-import org.ehcache.core.config.store.StoreConfigurationImpl;
+import org.ehcache.core.internal.store.StoreConfigurationImpl;
 import org.ehcache.config.units.MemoryUnit;
 import org.ehcache.core.spi.store.Store;
 import org.ehcache.expiry.Expiry;
